+++ R/globals.R
@@ -1,6 +1,7 @@
 #' @importFrom data.table := data.table frank setorder setnames
 #' @importFrom stats cor median sd var quantile setNames
 #' @importFrom Matrix sparseMatrix t
+#' @importFrom methods as
 NULL
 
 utils::globalVariables(
@@ -15,6 +16,7 @@ utils::globalVariables(
     "pair_label", "score_type", "score_value",
     "dim1", "dim2", "x", "y", "gene", "spatial_lee_p",
     "smoothed_cor", "neighbourhood_score", "cluster_cor",
+    "cross_celltype_score",
     "smoothed_expr1", "smoothed_expr2", "expression", "expr_status",
-    "metric", "metric_value")
+    "metric", "metric_value", "type")
 )